[#986] fix unmountComponentAtNode()
```
raven.js:3801 Warning: unmountComponentAtNode(): The node you're attempting to unmount was rendered by another copy of React.
raven.js:3801 Warning: render(...): It looks like the React-rendered content of this container was removed without using React. This is not supported and will cause errors. Instead, call ReactDOM.unmountComponentAtNode to empty a container.
raven.js:4023 Uncaught DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
```